refactor(tests): move test utilities to core/testing following Angular best practices
- Move test-utils.ts from app/test/ to app/core/testing/
- Add index.ts barrel export for cleaner imports
- Update all imports to use new location
- Remove test type from ESLint boundaries configuration
- Update CLAUDE.md documentation with new location

This change aligns with Angular's architecture pattern (@angular/core/testing)
and Thomas Trajan's enterprise architecture recommendations where testing
infrastructure belongs in the core module.
