+++ frontend/projects/webapp/src/app/core/testing/index.ts
@@ -1,0 +1,1 @@
+export * from './test-utils';
