+++ frontend/projects/webapp/src/app/feature/budget/create/services/template-selection.spec.ts
@@ -5,7 +5,7 @@ import { vi, describe, it, expect, beforeEach } from 'vitest';
 import { TemplateSelection } from './template-selection';
 import { TemplateApi } from '../../../../core/template/template-api';
 import { type TemplateLine, type BudgetTemplate } from '@pulpe/shared';
-import { createMockResourceRef } from '../../../../test/test-utils';
+import { createMockResourceRef } from '../../../../core/testing';
 
 // Mock interfaces for tests
 interface TemplateTotals {
